Fix broken win32 build
MinGW doesn't think a gsize is an unsigned long.

git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@23377 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/gnome-utils/gnc-tree-view.c
@@ -826,7 +826,7 @@ gnc_tree_view_get_column_order (GncTreeView *view,
 {
     const GList *tmp;
     GList *columns;
-    gsize num_cols = 0;
+    gulong num_cols = 0;
     gchar *col_names = NULL;
     gchar **col_str_list;
